Reworked User edit form to include roles property.
FYI, there's a problem in `nxus-web` with the handling of model
attributes and display fields. Model attribute definitions are collected
by the `ViewController` `_modelAttributes()` method. It uses the
`displayFields` property, if defined, to filter the attribute list. When
the attribute definitions are used in a context other than the entity
display (say the list view or an edit form) that requires a different
set of field definitions, required definitiony may be missing. For the
time being, we're just stuffing all the fields we use in any context
into displayFields.